# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elgiganten Repair API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/fabrikanter": {
            "get": {
                "tags": [
                    "Produkter - Fabrikanter"
                ],
                "summary": "Liste af Fabrikanter",
                "description": "",
                "operationId": "fabrikantsList",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "navn": {
                                            "type": "string",
                                            "example": "Bedste Fabrikant Navn"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Fabrikant Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    },
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Fabrikant Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Fabrikanter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produkter - Fabrikanter"
                ],
                "summary": "Opret en ny Fabrikant",
                "description": "",
                "operationId": "fabrikantsCreate",
                "parameters": [
                    {
                        "name": "navn",
                        "in": "query",
                        "description": "Fabrikant navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Fabrikant oprettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/fabrikanter/{id}": {
            "get": {
                "tags": [
                    "Produkter - Fabrikanter"
                ],
                "summary": "Find en specifik Fabrikant",
                "description": "",
                "operationId": "fabrikantsSpecific",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fabrikant ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "navn": {
                                            "type": "string",
                                            "example": "Bedste Fabrikant Navn"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Fabrikant Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Fabrikant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produkter - Fabrikanter"
                ],
                "summary": "Opdater en Fabrikant",
                "description": "",
                "operationId": "fabrikantsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fabrikant ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "navn",
                        "in": "query",
                        "description": "Fabrikant navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fabrikant Opdateret"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produkter - Fabrikanter"
                ],
                "summary": "Slet en Fabrikant",
                "description": "",
                "operationId": "fabrikantsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fabrikant ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Fabrikant Slettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}